+++ app/main/sequences.py
@@ -5,6 +5,7 @@ import logging
 from flask import Flask, redirect, render_template, request, session, abort
 from . import main
 from app.model import db, Sequence
+import re
 
 @main.route('/save_sequence', methods=['POST'])
 def save_sequence():
@@ -13,6 +14,10 @@ def save_sequence():
     else:
         seq_name = request.form.get("seq_name")
         seq_data = request.form.get("seq_data")
+        if re.match(r"^$|\s+", seq_name):
+            return "Un nom de séquence ne doit pas être vide ou contenir d'espace.", 400
+        if seq_data==None:
+            return "La séquence est vide.", 400
         logging.info("Saving sequence "+seq_name)
         db_sequence = Sequence(id=seq_name, value=seq_data, enabled=True)
         db.session.merge(db_sequence)
@@ -25,6 +30,8 @@ def enable_sequence():
         return render_template('login.html')
     else:
         seq_name = request.form.get("seq_name")
+        if re.match(r"^$|\s+", seq_name):
+            return "Un nom de séquence ne doit pas être vide ou contenir d'espace.", 400
         logging.info("Updating "+seq_name)
         db_seq = Sequence.query.filter_by(id=seq_name).first()
         db_seq.enabled = not db_seq.enabled
@@ -37,6 +44,8 @@ def delete_sequence():
         return render_template('login.html')
     else:
         seq_name = request.form.get("seq_name")
+        if re.match(r"^$|\s+", seq_name):
+            return "Un nom de séquence ne doit pas être vide ou contenir d'espace.", 400
         logging.info("Deleting "+seq_name)
         db_seq = Sequence.query.filter_by(id=seq_name).first()
         db.session.delete(db_seq)
